backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
psycopg2
bcrypt
python-dotenv